requirements.txt
flask==3.0.2  
python-dotenv==1.0.0  
pytest==8.3.5  
requests==2.32.3
certifi==2025.1.31
charset-normalizer==3.4.1
hvac==2.3.0
idna==3.10
requests==2.32.3
urllib3==2.3.0
wheel==0.45.1
